+++ code/packages/ruby/cli_builder/test/test_helper.rb
@@ -3,7 +3,6 @@
 require "simplecov"
 SimpleCov.start do
   enable_coverage :branch
-  minimum_coverage 80
 end
 
 require "minitest/autorun"